feat/training: allow relaxing the shape constraints for fit_generator
Some models need to reduce or more generally affect the size of a batch,
for instance because they require auxiliary samples to compute the
output for some target samples (e.g. graph convolutions which require
the neighbours of a batch to compute the output for that batch), or
because they compute an output that is differently shaped (e.g. an
adjacency matrix, when the input is a list of node features). In these
cases it can be useful to relax constraints on the shape of the training
arrays. Using `fit_generator(check_array_lengths=False)` allows exactly
that.